Prototype

+++ retesteth/RPCSession.cpp
@@ -39,181 +39,85 @@
 using namespace std;
 using namespace dev;
 
-IPCSocket::IPCSocket(string const& _path): m_path(_path)
-{
-#if defined(_WIN32)
-	m_socket = CreateFile(
-		m_path.c_str(),   // pipe name
-		GENERIC_READ |  // read and write access
-		GENERIC_WRITE,
-		0,              // no sharing
-		NULL,           // default security attribute
-		OPEN_EXISTING,  // opens existing pipe
-		0,              // default attributes
-		NULL);          // no template file
-
-	if (m_socket == INVALID_HANDLE_VALUE)
-		ETH_FAIL("Error creating IPC socket object!");
-
-#else
-	if (_path.length() >= sizeof(sockaddr_un::sun_path))
-		ETH_FAIL("Error opening IPC: socket path is too long!");
-
-	struct sockaddr_un saun;
-	memset(&saun, 0, sizeof(sockaddr_un));
-	saun.sun_family = AF_UNIX;
-	strcpy(saun.sun_path, _path.c_str());
-
-// http://idletechnology.blogspot.ca/2011/12/unix-domain-sockets-on-osx.html
-//
-// SUN_LEN() might be optimal, but it seemingly affects the portability,
-// with at least Android missing this macro.  Just using the sizeof() for
-// structure seemingly works, and would only have the side-effect of
-// sending larger-than-required packets over the socket.  Given that this
-// code is only used for unit-tests, that approach seems simpler.
-#if defined(__APPLE__)
-	saun.sun_len = sizeof(struct sockaddr_un);
-#endif //  defined(__APPLE__)
-
-	if ((m_socket = socket(AF_UNIX, SOCK_STREAM, 0)) < 0)
-		ETH_FAIL("Error creating IPC socket object");
-
-	if (connect(m_socket, reinterpret_cast<struct sockaddr const*>(&saun), sizeof(struct sockaddr_un)) < 0)
-	{
-		close(m_socket);
-		ETH_FAIL("Error connecting to IPC socket: " + _path);
-	}
-#endif
-}
-
-string IPCSocket::sendRequest(string const& _req)
-{
-#if defined(_WIN32)
-	// Write to the pipe.
-	DWORD cbWritten;
-	BOOL fSuccess = WriteFile(
-		m_socket,               // pipe handle
-		_req.c_str(),           // message
-		_req.size(),            // message length
-		&cbWritten,             // bytes written
-		NULL);                  // not overlapped
-
-	if (!fSuccess || (_req.size() != cbWritten))
-		ETH_FAIL("WriteFile to pipe failed");
-
-	// Read from the pipe.
-	DWORD cbRead;
-	fSuccess = ReadFile(
-		m_socket,          // pipe handle
-		m_readBuf,         // buffer to receive reply
-		sizeof(m_readBuf), // size of buffer
-		&cbRead,           // number of bytes read
-		NULL);             // not overlapped
-
-	if (!fSuccess)
-		ETH_FAIL("ReadFile from pipe failed");
-
-	return string(m_readBuf, m_readBuf + cbRead);
-#else
-	if (send(m_socket, _req.c_str(), _req.length(), 0) != (ssize_t)_req.length())
-		ETH_FAIL("Writing on IPC failed.");
-
-	auto start = chrono::steady_clock::now();
-	ssize_t ret;
-	string reply;
-	do
-	{
-		ret = recv(m_socket, m_readBuf, sizeof(m_readBuf), 0);
-
-		// Also consider closed socket an error.
-		if (ret < 0)
-			ETH_FAIL("Reading on IPC failed.");
-
-        // check for a long message
-        if (ret != 0)
-		{
-			ssize_t ret2 = ret;
-			do
-			{
-				reply += string(m_readBuf, m_readBuf + ret2);
-				ret2 = recv(m_socket, m_readBuf, sizeof(m_readBuf), MSG_DONTWAIT);
-				if (ret2 > 0)
-					std::this_thread::sleep_for(std::chrono::milliseconds(100)); //wait for socket update
-			}
-			while (ret2 > 0);
-		}
-	}
-	while (
-		ret == 0 &&
-		chrono::duration_cast<chrono::milliseconds>(chrono::steady_clock::now() - start).count() < m_readTimeOutMS
-	);
-
-	if (ret == 0)
-		ETH_FAIL("Timeout reading on IPC.");
-
-	return reply;
-#endif
-}
-
 struct sessionInfo
 {
-    sessionInfo(FILE* _pipe, RPCSession* _session, std::string const& _tmpDir, int _pid)
+    sessionInfo(
+        FILE* _pipe, RPCSession* _session, std::string const& _tmpDir, int _pid, unsigned _configId)
     {
         session.reset(_session);
         filePipe.reset(_pipe);
         tmpDir = _tmpDir;
         pipePid = _pid;
         isUsed = RPCSession::NotExist;
+        configId = _configId;
     }
     std::unique_ptr<RPCSession> session;
     std::unique_ptr<FILE> filePipe;
     int pipePid;
     RPCSession::SessionStatus isUsed;
     std::string tmpDir;
+    unsigned configId;
 };
+
+void closeSession(const string& _threadID);
 
 std::mutex g_socketMapMutex;
 static std::map<std::string, sessionInfo> socketMap;
-void RPCSession::runNewInstanceOfAClient(string const& _threadID)
+void RPCSession::runNewInstanceOfAClient(string const& _threadID, ClientConfig const& _config)
 {
-    fs::path tmpDir = test::createUniqueTmpDirectory();
-    string ipcPath = tmpDir.string() + "/geth.ipc";
-
-    string command = "eth";
-    std::vector<string> args;
-    args.push_back("--test");           // 1
-    args.push_back("--db-path");        // 2
-    args.push_back(tmpDir.string());    // 3
-    args.push_back("--ipcpath");        // 4
-    args.push_back(ipcPath);            // 5
-    args.push_back("--log-verbosity");  // 6
-    args.push_back("5");                // 7
-
-    int pid = 0;
-    test::popenOutput mode = (Options::get().enableClientsOutput) ? test::popenOutput::EnableALL :
-                                                                    test::popenOutput::DisableAll;
-    FILE* fp = test::popen2(command, args, "r", pid, mode);
-    if (!fp)
+    if (_config.getType() == Socket::IPC)
     {
-        ETH_ERROR("Failed to start the client: '" + command + "'");
-        std::raise(SIGABRT);
+        fs::path tmpDir = test::createUniqueTmpDirectory();
+        string ipcPath = tmpDir.string() + "/geth.ipc";
+
+        string command = "bash";
+        std::vector<string> args;
+        args.push_back(_config.getShellPath().c_str());
+        args.push_back(tmpDir.string());
+        args.push_back(ipcPath);
+
+        int pid = 0;
+        test::popenOutput mode = (Options::get().enableClientsOutput) ?
+                                     test::popenOutput::EnableALL :
+                                     test::popenOutput::DisableAll;
+        FILE* fp = test::popen2(command, args, "r", pid, mode);
+        if (!fp)
+        {
+            ETH_ERROR("Failed to start the client: '" + command + "'");
+            std::raise(SIGABRT);
+        }
+        else
+        {
+            int maxSeconds = 25;
+            while (!boost::filesystem::exists(ipcPath) && maxSeconds-- > 0)
+                std::this_thread::sleep_for(std::chrono::seconds(1));
+            ETH_REQUIRE_MESSAGE(maxSeconds > 0, "Client took too long to start ipc!");
+            // Client has opened ipc socket. wait for it to initialize
+            std::this_thread::sleep_for(std::chrono::seconds(4));
+        }
+        // sessionInfo info(fp,
+        //    new RPCSession(Socket::SocketType::IPC, "/home/wins/.ethereum/geth.ipc"),
+        //    tmpDir.string(), pid, _config.getId());
+        sessionInfo info(fp, new RPCSession(Socket::SocketType::IPC, ipcPath), tmpDir.string(), pid,
+            _config.getId());
+        {
+            std::lock_guard<std::mutex> lock(
+                g_socketMapMutex);  // function must be called from lock
+            socketMap.insert(std::pair<string, sessionInfo>(_threadID, std::move(info)));
+        }
+    }
+    else if (_config.getType() == Socket::TCP)
+    {
+        sessionInfo info(NULL, new RPCSession(Socket::SocketType::TCP, _config.getAddress()), "", 0,
+            _config.getId());
+        {
+            std::lock_guard<std::mutex> lock(
+                g_socketMapMutex);  // function must be called from lock
+            socketMap.insert(std::pair<string, sessionInfo>(_threadID, std::move(info)));
+        }
     }
     else
-    {
-        int maxSeconds = 25;
-        while (!boost::filesystem::exists(ipcPath) && maxSeconds-- > 0)
-            std::this_thread::sleep_for(std::chrono::seconds(1));
-        ETH_REQUIRE_MESSAGE(maxSeconds > 0, "Client took too long to start ipc!");
-        // Client has opened ipc socket. wait for it to initialize
-        std::this_thread::sleep_for(std::chrono::seconds(4));
-    }
-
-    sessionInfo info(fp, new RPCSession(ipcPath), tmpDir.string(), pid);
-    // sessionInfo info(fp, new RPCSession("/home/wins/.ethereum/geth.ipc"), tmpDir.string(), pid);
-    {
-        std::lock_guard<std::mutex> lock(g_socketMapMutex);  // function must be called from lock
-        socketMap.insert(std::pair<string, sessionInfo>(_threadID, std::move(info)));
-    }
+        ETH_FAIL("Unknown Socket Type in runNewInstanceOfAClient");
 }
 
 RPCSession& RPCSession::instance(const string& _threadID)
@@ -221,25 +125,34 @@ RPCSession& RPCSession::instance(const string& _threadID)
     bool needToCreateNew = false;
     {
         std::lock_guard<std::mutex> lock(g_socketMapMutex);
+        unsigned currentConfigId = Options::getDynamicOptions().getCurrentConfig().getId();
+        if (socketMap.count(_threadID) && socketMap.at(_threadID).configId != currentConfigId)
+        {
+            // For this thread a session is opened but it is opened not for current tested client
+            ETH_FAIL("A session opened for another client id!");
+        }
+
         if (!socketMap.count(_threadID))
         {
             // look for free clients that already instantiated
             for (auto& socket : socketMap)
             {
                 if (socket.second.isUsed == SessionStatus::Available)
-                {
-                    socket.second.isUsed = SessionStatus::Working;
-                    socketMap.insert(
-                        std::pair<string, sessionInfo>(_threadID, std::move(socket.second)));
-                    socketMap.erase(socketMap.find(socket.first));
-                    return *(socketMap.at(_threadID).session.get());
+                    if (socket.second.configId == currentConfigId)
+                    {
+                        socket.second.isUsed = SessionStatus::Working;
+                        socketMap.insert(
+                            std::pair<string, sessionInfo>(_threadID, std::move(socket.second)));
+                        socketMap.erase(socketMap.find(socket.first));  // remove previous threadID
+                                                                        // assigment to this socket
+                        return *(socketMap.at(_threadID).session.get());
                 }
             }
             needToCreateNew = true;
         }
     }
     if (needToCreateNew)
-        runNewInstanceOfAClient(_threadID);
+        runNewInstanceOfAClient(_threadID, Options::getDynamicOptions().getCurrentConfig());
     std::lock_guard<std::mutex> lock(g_socketMapMutex);
     ETH_REQUIRE_MESSAGE(socketMap.size() <= Options::get().threadCount,
         "Something went wrong. Retesteth create more instances than needed!");
@@ -272,13 +185,16 @@ RPCSession::SessionStatus RPCSession::sessionStatus(std::string const& _threadID
 
 void closeSession(const string& _threadID)
 {
-    ETH_REQUIRE(socketMap.count(_threadID));
+    ETH_REQUIRE_MESSAGE(socketMap.count(_threadID), "Socket map is empty in closeSession!");
     sessionInfo& element = socketMap.at(_threadID);
-    test::pclose2(element.filePipe.get(), element.pipePid);
-    std::this_thread::sleep_for(std::chrono::seconds(4));
-    boost::filesystem::remove_all(boost::filesystem::path(element.tmpDir));
-    element.filePipe.release();
-    element.session.release();
+    if (element.session.get()->getSocketType() == Socket::SocketType::IPC)
+    {
+        test::pclose2(element.filePipe.get(), element.pipePid);
+        std::this_thread::sleep_for(std::chrono::seconds(4));
+        boost::filesystem::remove_all(boost::filesystem::path(element.tmpDir));
+        element.filePipe.release();
+        element.session.release();
+    }
 }
 
 void RPCSession::clear()
@@ -349,7 +265,7 @@ string RPCSession::eth_sendTransaction(string const& _transaction)
 
 string RPCSession::eth_sendRawTransaction(std::string const& _rlp)
 {
-	return rpcCall("eth_sendRawTransaction", { quote(_rlp) }).asString();
+    return rpcCall("eth_sendRawTransaction", {quote(_rlp)}, true).asString();
 }
 
 std::string RPCSession::eth_getTransactionCount(std::string const& _address, std::string const& _blockNumber)
@@ -421,7 +337,7 @@ void RPCSession::test_setChainParams(vector<string> const& _accounts)
 	)";
 
 	Json::Value config;
-    ETH_REQUIRE(Json::Reader().parse(c_configString, config));
+    ETH_REQUIRE_MESSAGE(Json::Reader().parse(c_configString, config), "setChainParams json is incorrect!");
 	for (auto const& account: _accounts)
 		config["accounts"][account]["wei"] = "0x100000000000000000000000000000000000000000";
 	test_setChainParams(Json::FastWriter().write(config));
@@ -444,18 +360,19 @@ void RPCSession::test_importRawBlock(std::string const& _blockRLP)
 
 void RPCSession::test_setChainParams(string const& _config)
 {
-    ETH_REQUIRE(rpcCall("test_setChainParams", { _config }) == true);
+    ETH_REQUIRE_MESSAGE(rpcCall("test_setChainParams", { _config }) == true, "remote test_setChainParams = false");
 }
 
 void RPCSession::test_rewindToBlock(size_t _blockNr)
 {
-    ETH_REQUIRE(rpcCall("test_rewindToBlock", { to_string(_blockNr) }) == true);
+    ETH_REQUIRE_MESSAGE(rpcCall("test_rewindToBlock", { to_string(_blockNr) }) == true, "remote test_rewintToBlock = false");
 }
 
 void RPCSession::test_mineBlocks(int _number, string const& _hash)
 {
-    // u256 startBlock = fromBigEndian<u256>(fromHex(rpcCall("eth_blockNumber").asString()));
-    ETH_REQUIRE(rpcCall("test_mineBlocks", { to_string(_number) }, true) == true);
+       (void)_hash;
+    u256 startBlock = fromBigEndian<u256>(fromHex(rpcCall("eth_blockNumber").asString()));
+    ETH_REQUIRE_MESSAGE(rpcCall("test_mineBlocks", { to_string(_number) }, true) == true, "remote test_mineBlocks = false");
 
 	// We auto-calibrate the time it takes to mine the transaction.
 	// It would be better to go without polling, but that would probably need a change to the test client
@@ -473,11 +390,11 @@ void RPCSession::test_mineBlocks(int _number, string const& _hash)
 			//ETH_FAIL("Error in test_mineBlocks: block mining timeout! " + test::TestOutputHelper::get().testName());
 
         // std::cerr << test_getBlockStatus(_hash) << std::endl;
-        // bigint number = fromBigEndian<u256>(fromHex(rpcCall("eth_blockNumber").asString()));
-        // if (number >= startBlock + _number)
-        //	break;
-        if (test_getBlockStatus(_hash) == "Ready")
+         bigint number = fromBigEndian<u256>(fromHex(rpcCall("eth_blockNumber").asString()));
+         if (number >= startBlock + _number)
             break;
+        //if (test_getBlockStatus(_hash) == "Ready")
+        //    break;
         else
 			sleepTime *= 2;
 	}
@@ -500,39 +417,44 @@ void RPCSession::test_mineBlocks(int _number, string const& _hash)
 
 void RPCSession::test_modifyTimestamp(size_t _timestamp)
 {
-    ETH_REQUIRE(rpcCall("test_modifyTimestamp", { to_string(_timestamp) }) == true);
+    ETH_REQUIRE_MESSAGE(rpcCall("test_modifyTimestamp", { to_string(_timestamp) }) == true, "test_modifyTimestamp was not successfull");
+}
+
+std::string RPCSession::sendRawRequest(string const& _request)
+{
+    return m_socket.sendRequest(_request);
 }
 
 Json::Value RPCSession::rpcCall(string const& _methodName, vector<string> const& _args, bool _canFail)
 {
-	string request = "{\"jsonrpc\":\"2.0\",\"method\":\"" + _methodName + "\",\"params\":[";
-	for (size_t i = 0; i < _args.size(); ++i)
-	{
-		request += _args[i];
-		if (i + 1 != _args.size())
-			request += ", ";
-	}
+    string request = "{\"jsonrpc\":\"2.0\",\"method\":\"" + _methodName + "\",\"params\":[";
+    for (size_t i = 0; i < _args.size(); ++i)
+    {
+        request += _args[i];
+        if (i + 1 != _args.size())
+            request += ", ";
+    }
 
-	request += "],\"id\":" + to_string(m_rpcSequence) + "}";
-	++m_rpcSequence;
+    request += "],\"id\":" + to_string(m_rpcSequence) + "}";
+    ++m_rpcSequence;
 
     ETH_TEST_MESSAGE("Request: " + request);
-	string reply = m_ipcSocket.sendRequest(request);
+    string reply = m_socket.sendRequest(request);
     ETH_TEST_MESSAGE("Reply: " + reply);
 
-	Json::Value result;
-    ETH_REQUIRE(Json::Reader().parse(reply, result, false));
+    Json::Value result;
+    ETH_REQUIRE_MESSAGE(Json::Reader().parse(reply, result, false), "error parsing json from remote response!");
 
-	if (result.isMember("error"))
-	{
-		if (_canFail)
-			return Json::Value();
+    if (result.isMember("error"))
+    {
+        if (_canFail)
+            return Json::Value();
 
-		ETH_FAIL("Error on JSON-RPC call (" + test::TestOutputHelper::get().testName() + "): "
-		 + result["error"]["message"].asString()
-		 + " Request: " + request);
-	}
-	return result["result"];
+        ETH_FAIL("Error on JSON-RPC call (" + test::TestOutputHelper::get().testName() + "): "
+         + result["error"]["message"].asString()
+         + " Request: " + request);
+    }
+    return result["result"];
 }
 
 string const& RPCSession::accountCreate()
@@ -549,8 +471,8 @@ string const& RPCSession::accountCreateIfNotExists(size_t _id)
 	return m_accounts[_id];
 }
 
-RPCSession::RPCSession(const string& _path):
-	m_ipcSocket(_path)
+RPCSession::RPCSession(Socket::SocketType _type, const string& _path):
+    m_socket(_type, _path)
 {
 	//accountCreate();
 	//This will pre-fund the accounts create prior.
